Add a check here to fix the generation failed in the parallel execution

+++ hooks/01_removeNotRelevantParts.js
@@ -6,7 +6,6 @@ const callback = (error) => {
     throw error;
   }
 };
-
 /**
  * Removes unnecessary files (css, js) if user pass `singleFile` parameter. 
  */
@@ -21,8 +20,15 @@ module.exports = {
         maxBusyTries: 3
       };
 
-      rimraf.sync(jsDir, opts, callback);
-      rimraf.sync(cssDir, opts, callback);
+      // Attempt to delete jsDir if it exists
+      if (fs.existsSync(jsDir)) {
+        rimraf.sync(jsDir, opts, callback);
+      }
+
+      // Attempt to delete cssDir if it exists
+      if (fs.existsSync(cssDir)) {
+        rimraf.sync(cssDir, opts, callback);
+      }
     }
   }
 };
